rbd: generalize schedule reusable code
